[cli] orgainize cli & create a wiki
